Pipeline/pyproject.toml
[tool.poetry]
name = "pipeline"
version = "0.1.0"
description = ""
authors = ["FlameInBrainSerj <Krivosheev.sergey.01@mail.ru>"]
readme = "README.md"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu118"
priority = "supplemental"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.0"
numpy = "^1.26.4"
hydra-core = "^1.3.2"
pyarrow = "^15.0.0"
fastparquet = "^2024.2.0"
natasha = "^1.6.0"
nltk = "^3.8.1"
matplotlib = "^3.8.3"
argparse = "^1.4.0"
torch = "^2.2.1"
torchtext = "^0.17.1"
torchvision = "^0.17.1"
pytorch-lightning = "^2.2.0.post0"
scikit-learn = "^1.4.1.post1"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"
flake8-bugbear = "^24.2.6"
flake8-builtins = "^2.2.0"
pep8-naming = "^0.13.3"
flake8-commas = "^2.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
